proj/requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Django==2.1
django-widget-tweaks==1.4.8
filetype==1.0.7
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
imutils==0.5.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.4
pdf2image==1.14.0
Pillow==8.0.1
plotly==4.12.0
poppler-utils==0.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
requests==2.25.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scipy==1.4.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-hub==0.8.0
termcolor==1.1.0
tifffile==2020.10.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
